655-final-project-backend/cloud_functions_be_microservices/final-project-655-add-url-image-into-gcs-bucket/requirements.txt
functions-framework==3.*
google-cloud-storage
requests
pillow
protobuf
jsonify
